Modularise project structure and update documentation
