+++ packages/cdk/constructs/LambdaFunction.ts
@@ -132,7 +132,7 @@ export class LambdaFunction extends Construct {
       memorySize: 256,
       timeout: Duration.seconds(50),
       architecture: Architecture.X86_64,
-      handler: "app.handler",
+      handler: "app.handler.handler",
       code: Code.fromAsset(props.packageBasePath),
       role,
       environment: {
