+++ packages/syncKnowledgeBaseFunction/app/handler.py
@@ -1,40 +1,25 @@
-import os
 import time
 import boto3
 from botocore.exceptions import ClientError
 from aws_lambda_powertools import Logger
+from app.config.config import KNOWLEDGEBASE_ID, DATA_SOURCE_ID
 
-
-# Initialize Powertools Logger with service name for better log organization
 logger = Logger(service="syncKnowledgeBaseFunction")
-
-# Initialize Bedrock client for knowledge base operations
 bedrock_agent = boto3.client("bedrock-agent")
 
 
 @logger.inject_lambda_context
 def handler(event, context):
-    """
-    Lambda handler that processes S3 events and triggers Bedrock Knowledge Base ingestion.
-
-    This function is triggered when documents are created, updated, deleted, or restored in the S3 bucket
-    and automatically starts an ingestion job to sync the knowledge base with current content.
-    """
-    # Record start time for performance tracking
+    """Lambda handler that processes S3 events and triggers Bedrock Knowledge Base ingestion."""
     start_time = time.time()
 
-    # Get required environment variables
-    knowledge_base_id = os.environ.get("KNOWLEDGEBASE_ID")
-    data_source_id = os.environ.get("DATA_SOURCE_ID")
-
-    # Validate configuration
-    if not knowledge_base_id or not data_source_id:
+    if not KNOWLEDGEBASE_ID or not DATA_SOURCE_ID:
         logger.error(
             "Missing required environment variables",
             extra={
                 "status_code": 500,
-                "knowledge_base_id": bool(knowledge_base_id),
-                "data_source_id": bool(data_source_id),
+                "knowledge_base_id": bool(KNOWLEDGEBASE_ID),
+                "data_source_id": bool(DATA_SOURCE_ID),
             },
         )
         return {"statusCode": 500, "body": "Configuration error"}
@@ -42,8 +27,8 @@ def handler(event, context):
     logger.info(
         "Starting knowledge base sync process",
         extra={
-            "knowledge_base_id": knowledge_base_id,
-            "data_source_id": data_source_id,
+            "knowledge_base_id": KNOWLEDGEBASE_ID,
+            "data_source_id": DATA_SOURCE_ID,
         },
     )
 
@@ -51,10 +36,8 @@ def handler(event, context):
         processed_files = []
         job_ids = []
 
-        # Process each S3 event record
         for record_index, record in enumerate(event.get("Records", [])):
             if record.get("eventSource") == "aws:s3":
-                # Validate S3 event structure
                 s3_info = record.get("s3", {})
                 bucket_name = s3_info.get("bucket", {}).get("name")
                 object_key = s3_info.get("object", {}).get("key")
@@ -70,7 +53,6 @@ def handler(event, context):
                     )
                     continue
 
-                # Extract S3 event details
                 bucket = bucket_name
                 key = object_key
                 event_name = record["eventName"]
@@ -88,16 +70,14 @@ def handler(event, context):
                     },
                 )
 
-                # Start ingestion job for the knowledge base (handles all event types)
                 ingestion_start_time = time.time()
                 response = bedrock_agent.start_ingestion_job(
-                    knowledgeBaseId=knowledge_base_id,
-                    dataSourceId=data_source_id,
+                    knowledgeBaseId=KNOWLEDGEBASE_ID,
+                    dataSourceId=DATA_SOURCE_ID,
                     description=f"Auto-sync triggered by S3 {event_name} on {key}",
                 )
                 ingestion_request_time = time.time() - ingestion_start_time
 
-                # Extract job information
                 job_id = response["ingestionJob"]["ingestionJobId"]
                 job_status = response["ingestionJob"]["status"]
 
@@ -106,14 +86,13 @@ def handler(event, context):
                     extra={
                         "job_id": job_id,
                         "job_status": job_status,
-                        "knowledge_base_id": knowledge_base_id,
+                        "knowledge_base_id": KNOWLEDGEBASE_ID,
                         "trigger_file": key,
                         "ingestion_request_duration_ms": round(ingestion_request_time * 1000, 2),
                         "note": "Job will process all files in data source, not just trigger file",
                     },
                 )
 
-                # Track processed files and job IDs for summary
                 processed_files.append(key)
                 job_ids.append(job_id)
             else:
@@ -125,10 +104,8 @@ def handler(event, context):
                     },
                 )
 
-        # Calculate total processing time
         total_duration = time.time() - start_time
 
-        # Log successful completion summary
         logger.info(
             "Knowledge base sync process completed",
             extra={
@@ -138,7 +115,7 @@ def handler(event, context):
                 "job_ids": job_ids,
                 "trigger_files": processed_files,
                 "total_duration_ms": round(total_duration * 1000, 2),
-                "knowledge_base_id": knowledge_base_id,
+                "knowledge_base_id": KNOWLEDGEBASE_ID,
                 "next_steps": "Monitor Bedrock console for ingestion job completion status",
             },
         )
@@ -151,11 +128,9 @@ def handler(event, context):
         }
 
     except ClientError as e:
-        # Handle AWS service errors with detailed logging
         error_code = e.response.get("Error", {}).get("Code", "Unknown")
         error_message = e.response.get("Error", {}).get("Message", str(e))
 
-        # Handling for ConflictException
         if error_code == "ConflictException":
             logger.warning(
                 "Ingestion job already in progress - no action required",
@@ -165,8 +140,8 @@ def handler(event, context):
                     "error_message": error_message,
                     "description": "Files uploaded successfully and will be processed by existing ingestion job",
                     "action_required": "none",
-                    "knowledge_base_id": knowledge_base_id,
-                    "data_source_id": data_source_id,
+                    "knowledge_base_id": KNOWLEDGEBASE_ID,
+                    "data_source_id": DATA_SOURCE_ID,
                     "duration_ms": round((time.time() - start_time) * 1000, 2),
                     "explanation": (
                         "This is normal when multiple files are uploaded quickly. "
@@ -185,8 +160,8 @@ def handler(event, context):
                     "status_code": 500,
                     "error_code": error_code,
                     "error_message": error_message,
-                    "knowledge_base_id": knowledge_base_id,
-                    "data_source_id": data_source_id,
+                    "knowledge_base_id": KNOWLEDGEBASE_ID,
+                    "data_source_id": DATA_SOURCE_ID,
                     "duration_ms": round((time.time() - start_time) * 1000, 2),
                 },
             )
@@ -196,7 +171,6 @@ def handler(event, context):
             }
 
     except Exception as e:
-        # Handle unexpected errors
         logger.error(
             "Unexpected error occurred",
             extra={
